+++ CHANGES.rst
@@ -3,7 +3,7 @@ Changelog
 
 There's a frood who really knows where his towel is.
 
-1.0b2 (unreleased)
+1.0b2 (2018-01-29)
 ------------------
 
 - Remove indirect dependency on five.grok.